CRT_Buddy/requirements.txt
PyQt6==6.6.1
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
pygame==2.5.2
pywin32==306
requests==2.31.0
